Fondos con velocidad del nivel. Clouds aplicado a un nivel esta vez posta de verdad

+++ classes/model/CampoMovil.h
@@ -29,7 +29,7 @@ private:
 	int ancho;
 	int alto;
 	Vector posicion;
-	int velocidadX;
+	float velocidadX;
 	Jugador* jugador;
 	CampoVista* vista;
 	std::list<Entidad*> entidades;
